Improve xposedminversion handling
Allow using "2.1+" as minimum version instead of just "2.1*". I think this
simply looks better in this context. :)

Don't crash when the xposedminversion metadata exists, but in the wrong format (e.g.
when the value is "2.1" instead of "2.1+", it is saved as float).

Don't trim trailing ".0"s from displayed versions. "2.0" looks better than
just "2" in my opinion.