XposedBridge v32, non-static BusyBox executables
This reduces the file size from 624 kB to 366 kB. Even if storage space
isn't a problem, this avoids unnecessary traffic (183k downloads for
Xposed 2.2 so far in less than a month, so the saved traffic adds up to
almost 50 GB, which could easily become more with more frequent updates
and more users).

It also shouldn't be necessary to statically link libc etc. into BusyBox,
this isn't the case for the versions that come with CyanogenMod and the
Genymotion emulator either. This drastically reduces the file size.