Installation via recovery
With this, the Xposed app_process binary can be installed via a custom
recovery (CWM or TWRP). This is especially use full for HTC phones with
S-On, which reverts any changes to the /system partition on reboot.

The flashable zip contains shell script which creates the backup and then
copies the new app_process file from this location:
/data/data/de.robv.android.xposed.installer/bin/app_process.
Therefore, it is independent from the Xposed version and platform, doesn't
need to be generated dynamically and could also be used to restore Xposed
after ROM upgrades.

Still missing:
- GUI/settings to use the classical installation method as default and
  give the user a hint that he can switch to a different one in case of
  problems.
- Better summary when the installation via recovery has been prepared.
- Confirmation dialog before rebooting to recovery.
- More testing with different recoveries and devices.