Load static Busybox executables from separate package
This makes it possible to ship with a dynamically compiled Busybox for the
99.9% who don't have problems like "Segmentation fault", and offer a
separate download of the static versions (much bigger!) for both ARM and
x86 for those who do experience such issues.

There is currently no UI feedback on this, like showing that the static
version is used or offering the download (maybe even automatically if
typical errors are detected). This might be added later, depending on
the feedback received.

Note that the static package needs to be signed with the same key in order
to prevent attackers from creating a malicious executable which Xposed
Installer would start with root permissions.