Activate module from a notification action
After installing a new module there is a notification posted to the
notifcation bar telling the user to activate the module.

This commit adds two notification actions: "Activate and reboot" and
"Activate".

"Activate and reboot" activates the module and reboots the device (full
reboot).
"Activate" activates the module with no further action. This is usefull
if more modules are installed at once but most likely this action will
be used less then "Activate and reboot".

This is how it looks
http://i.imgur.com/A5uOR3a.png