Don't show all versions - add "More" card
Add a "More" card to the module versions list. Only load all versions if
necessary.

See a live demonstration here
http://youtu.be/bx8jL4T3iUE

If accepted, the "More" strings has to be translated as well. Since all
translations are updated now, this might be considered for the next
release.

The implementation is to be improved.