Update strings.xml
This is a more refined and enhanced version of the translation. I took the "strings.xml" file located in XposedInstaller-master\res\values , which is in English and translated it from scratch.
The "strings.xml" in Arabic - translated by @aljohanei - was translated from English using Google translator and contain many mistakes.
Of course me lacking the knowledge of how to but the file in the app and run it on my phone to validate ' Text Alignment ' , I couldn't verify if it works appropriately or not. But the translation is accurate.