Update strings.xml
This is a more refined and enhanced version of the translation. I took the "strings.xml" file located in XposedInstaller_2.6.1.apk\res\values , which is in English and translated it from scratch.
The "strings.xml" in Arabic - translated by @aljohanei - was translated from English using Google translator and contain many mistakes.
I have verified - On My Onepluse one - that it works appropriately, all lines have been translated and the translation is 100% accurate.