Use version number from /system/xposed.prop
The Xposed framework consists of several parts: app_process, XposedBridge.jar
and the modified ART runtime. These parts have been modified to use single
version number for the whole framework package instead of instead of having
different version numbers for the individual parts. It's not supported to mix
parts from different packages anyway.

As a result of this, there is no more need to determine the versions of
app_process and XposedBridge.jar in the installer. It will be read from
the file /system/xposed.prop instead, which is copied to the device during
the framework installation.