Enable copying text from Logs screen
Currently, you can't press on the Xposed logs to copy some lines (show text processing panel). With this patch, we can select and copy what we want. We don't have to do it longer way: export the Xposed log, find the log file, open a text editor, select part of the logs to copy.